+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This version of Wgit see's a major refactor of the code base involving multiple 
 - `Wgit::Database.connect` method (alias for `Wgit::Database.new`).
 - `Wgit::Url#absolute?` method.
 - `Wgit::Url#relative? base: url` support.
+- `case_sensitive:` and `whole_sentence:` support to `Database#search` and `Document#search` methods.
 ### Changed/Removed
 - Breaking changes: Renamed the following `Wgit` and `Wgit::Indexer` methods: `Wgit.index_the_web` to `Wgit.index_www`, `Wgit::Indexer.index_the_web` to `Wgit::Indexer.index_www`, `Wgit.index_this_site` to `Wgit.index_site`, `Wgit::Indexer.index_this_site` to `Wgit::Indexer.index_site`, `Wgit.index_this_page` to `Wgit.index_page`, `Wgit::Indexer.index_this_page` to `Wgit::Indexer.index_page`.
 - Breaking changes: All `Wgit::Indexer` methods now take named parameters.
@@ -32,6 +33,7 @@ This version of Wgit see's a major refactor of the code base involving multiple 
 - Breaking changes: The following `Wgit::Assertable` method aliases have been removed: `.type`, `.types` (use `.assert_types` instead) and `.arr_type`, `.arr_types` (use `.assert_arr_types` instead).
 - Breaking changes: The following `Wgit::Utils` methods now take named parameters: `.to_h` and `.printf_search_results`.
 - `Wgit::Document` instances can now be instantiated with `String` Url's (previously only `Wgit::Url`'s).
+- `Wgit::Utils.printf_search_results`'s method signature has changed; the search parameters have been removed. Before calling this method you must call `doc.search!` on each of the `results`. See the docs for the full details.
 ### Fixed
 - ...
 ---